backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
opencv-python-headless
ultralytics
sahi
bcos
torch
torchvision
captum
lime
scikit-image